+++ src/helpers/StringHelper.php
@@ -1892,4 +1892,25 @@ class StringHelper extends \yii\helpers\StringHelper
             random_int(0, 0xffff), random_int(0, 0xffff), random_int(0, 0xffff)
         );
     }
+
+    /**
+     * Converts an email from IDNA ASCII to Unicode, if the Intl extension is installed.
+     *
+     * @param string $email
+     * @return string
+     * @since 3.5.16
+     */
+    public static function idnToUtf8Email(string $email): string
+    {
+        if (!function_exists('idn_to_utf8')) {
+            return $email;
+        }
+        $parts = explode('@', $email, 2);
+        foreach ($parts as &$part) {
+            if (($part = idn_to_utf8($part)) === false) {
+                return $email;
+            }
+        }
+        return implode('@', $parts);
+    }
 }